+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added a static function to create a `PickingHitArea` without the need to pass a renderer.
+
 ## [1.3.1] - 2022-03-18
 ### Fixed
 - Fixed an issue which caused skinning to work incorrectly when PIXI.settings.RESOLUTION = 2.